REFACTOR: rename components to widgets
