+++ lib/presentation/pages/schedule/widgets/lesson_card.dart
